tests/requirements.txt
black==18.9b0
coverage==4.5.2
devtools==0.5
docutils==0.14
flake8==3.6.0
isort==4.3.4
pycodestyle==2.4.0
pyflakes==2.0.0
pygments==2.2.0
pytest==4.0.0
pytest-aiohttp==0.3.0
pytest-cov==2.6.0
pytest-isort==0.2.1
pytest-mock==1.10.0
pytest-sugar==0.9.2
pytest-toolbox==0.4
